requirements.txt
Flask
Flask-Cors
flask-SQLAlchemy
psycopg2-binary
Flask-Migrate
Flask-Script
Flask-Mail
pytest
Flask-JWT-Extended
selenium
gunicorn
boto3